pyproject.toml
[tool.poetry]
name = "day1"
version = "0.1.0"
description = ""
authors = ["sohanur rahman <shohanur.rahman@gennext.net>"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
pydantic = "^2.8.2"
sqlalchemy = "^2.0.32"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"

[tool.poetry.scripts]
app = "src.main:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
